Changing the TrigTime from an unsigned int to an int64_t to accomodate a 64 bit timestamps being stored there.

+++ RawData/ExternalTrigger.cxx
@@ -21,7 +21,7 @@ namespace raw{
   }
 
   //----------------------------------------------------------------------
-  ExternalTrigger::ExternalTrigger(unsigned int trigid, unsigned int trigtime) 
+  ExternalTrigger::ExternalTrigger(unsigned int trigid, int64_t trigtime) 
     : fTrigID(trigid)
     , fTrigTime(trigtime)
   {
