+++ RawData/ExternalTrigger.h
@@ -12,6 +12,7 @@
 #include <vector>
 #include <iosfwd>
 #include <time.h>
+#include <stdint.h>
 
 namespace raw {
 
@@ -22,21 +23,21 @@ namespace raw {
     private:
 
       unsigned int   fTrigID;
-      unsigned int   fTrigTime;
+      int64_t        fTrigTime;
 
 #ifndef __GCCXML__
 
   public:
 
-      ExternalTrigger(unsigned int trigid, unsigned int trigtime);
+      ExternalTrigger(unsigned int trigid, int64_t trigtime);
 
       // Set Methods
       void             SetTrigID(unsigned int i);
-      void             SetTrigTime(unsigned int i);
+      void             SetTrigTime(int64_t i);
 
       // Get Methods
       unsigned int     GetTrigID()          const;
-      unsigned int     GetTrigTime()          const;
+      int64_t          GetTrigTime()          const;
      
 #endif
     };
@@ -45,9 +46,9 @@ namespace raw {
 #ifndef __GCCXML__
 
 inline void           raw::ExternalTrigger::SetTrigID(unsigned int i)    { fTrigID = i;      }
-inline void           raw::ExternalTrigger::SetTrigTime(unsigned int i)  { fTrigTime = i;    }
+inline void           raw::ExternalTrigger::SetTrigTime(int64_t i)       { fTrigTime = i;    }
 inline unsigned int   raw::ExternalTrigger::GetTrigID()           const  { return fTrigID;   }  
-inline unsigned int   raw::ExternalTrigger::GetTrigTime()         const  { return fTrigTime; }  
+inline int64_t        raw::ExternalTrigger::GetTrigTime()         const  { return fTrigTime; }  
 
 #endif
